- [ZBXNEXT-672] updated to latest trunk
git-svn-id: svn://svn.zabbix.com/branches/dev@18323 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ ZBXNEXT-672/frontends/php/api/classes/class.cwebcheck.php
@@ -498,7 +498,9 @@ COpt::memoryPick();
 				$itemids_del[] = $stepitem['itemid'];
 			}
 
-			API::Item()->delete($itemids_del, true);
+			if(!empty($itemids_del)){
+				API::Item()->delete($itemids_del, true);
+			}
 
 			DB::delete('httptest', array('httptestid' => $webcheckids));
 
@@ -760,7 +762,9 @@ COpt::memoryPick();
 
 		DB::delete('httpstep', array('httpstepid' => $webstepids));
 
-		API::Item()->delete($itemids, true);
+		if(!empty($itemids)){
+			API::Item()->delete($itemids, true);
+		}
 	}
 }
 ?>
